refactor: URI 수정 및 패키지 이동
