+++ README.md
@@ -4,12 +4,12 @@
 
 ### API 명세서
 
-| Method | URI                  | Description                 |
-|--------|----------------------|-----------------------------|
-| POST   | /lines/stations      | 노선에 역을 등록하는 API             |
-| DELETE | /lines/stations/{id} | 노선의 역을 삭제하는 API             |
-| GET    | /lines/{id}          | 노선의 역들을 조회하는 API            |
-| GET    | /lines               | 모든 노선과 노선에 포함된 역들을 조회하는 API |
+| Method | URI               | Description                 |
+|--------|-------------------|-----------------------------|
+| POST   | /lines/register   | 노선에 역을 등록하는 API             |
+| DELETE | /lines/unregister | 노선의 역을 삭제하는 API             |
+| GET    | /lines/{id}       | 노선의 역들을 조회하는 API            |
+| GET    | /lines            | 모든 노선과 노선에 포함된 역들을 조회하는 API |
 
 ### 기능 요구사항
